Use LifecycleManager instead of inheriting AbstractLifecycle
